requirements.txt
tensorflow-gpu==2.0.0-beta1
scikit-image
scikit-learn
pandas
mrcfile
